web/requirements/production.txt
amqp==2.1.4
appdirs==1.4.3
argh==0.26.2
asn1crypto==0.22.0
audioread==2.1.8
azure==1.0.3
azure-common==1.1.4
azure-mgmt==0.20.2
azure-mgmt-common==0.20.0
azure-mgmt-compute==0.20.1
azure-mgmt-network==0.20.1
azure-mgmt-nspkg==1.0.0
azure-mgmt-resource==0.20.1
azure-mgmt-storage==0.20.0
azure-nspkg==1.0.0
azure-servicebus==0.20.1
azure-servicemanagement-legacy==0.20.2
azure-storage==0.36.0
awscli==1.18.223
babelfish==0.5.5
backcall==0.1.0
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.8.2
billiard==3.5.0.2
boto==2.43.0
boto3==1.0.0
botocore==1.19.10
cachetools==4.0.0
celery==4.0.2
certifi==2019.11.28
cffi==1.10.0
chardet==3.0.4
click==6.7
cognitive-face==1.5.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==1.8.1
cycler==0.10.0
decorator==4.4.2
distro==1.4.0
dj-database-url==0.4.1
Django==1.11.17
django-admin-tools==0.9.0
django-advanced-filters==1.1.1
django-appconf==1.0.2
django-braces==1.14.0
django-bulk-update==2.2.0
django-celery==3.1.17
django-celery-results==1.0.1
django-configurations==2.2
django-cors-headers==2.0.0
django-crispy-forms==1.6.1
django-database-size==1.0.2
django-debug-toolbar==1.6
django-filter==1.0.4
django-grappelli==2.9.1
django-guardian==1.4.6
django-model-utils==4.0.0
django-mptt==0.8.6
django-nested-admin==3.2.4
django-nose==1.4.4
django-notifications-hq==1.5.0
django-python3-ldap==0.11.3
django-queued-storage==0.8.0
django-redis==4.6.0
django-rest-fuzzysearch==0.5.1
django-storages==1.5.2
django-versatileimagefield==2.0
djangorestframework==3.6.2
djoser==0.7.0
docutils==0.15.2
dogpile.cache==0.9.0
drf-compound-fields==1.0.0
drf-extensions==0.3.1
drf-nested-routers==0.11.1
drf-tracking==1.5.0
drfdocs==0.0.11
elasticsearch==7.6.0
enum34==1.1.6
enzyme==0.4.1
et-xmlfile==1.0.1
futures==2.2.0
gcloud==0.18.3
gevent==1.1.2
google-api-core==1.16.0
google-auth==1.11.3
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-language==1.3.0
google-cloud-speech==1.3.2
google-cloud-storage==1.26.0
google-cloud-translate==2.0.1
google-cloud-vision==1.0.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
greenlet==0.4.11
grpcio==1.0.4
guessit==3.1.0
gunicorn==19.6.0
httplib2==0.17.0
idna==2.5
imageio==2.8.0
imageio-ffmpeg==0.4.1
indic-transliteration==1.9.4
instagram-scraper==1.8.1
ipython==7.9.0
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jsonfield==2.1.1
kiwisolver==1.1.0
kombu==4.0.2
ldap3==2.8.1
librosa==0.7.2
llvmlite==0.31.0
lxml==4.5.0
Markdown==2.6.7
MarkupSafe==1.1.1
matplotlib==3.0.3
moviepy==1.0.1
nose==1.3.7
numba==0.47.0
numpy==1.18.2
oauth2client==4.1.3
olefile==0.44
opencv-contrib-python==3.3.0.10
opencv-python==4.2.0.32
openpyxl==2.6.4
packaging==16.8
pandas==0.24.2
parso==0.6.2
pathtools==0.1.2
pbr==5.4.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==4.0.0
proglog==0.1.9
prompt-toolkit==2.0.10
protobuf==3.11.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyacoustid==1.1.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.17
pycurl==7.43.0
pydub==0.23.1
Pygments==2.6.1
pygobject==3.20.0
pyparsing==2.2.0
PyPDF2==1.26.0
pysrt==1.1.2
python-apt==1.1.0b1
python-dateutil==2.6.0
python-docx==0.8.10
python-dotenv==0.12.0
python-ldap==3.2.0
python-magic==0.4.15
python-monkey-business==1.0.0
pytz==2016.7
PyYAML==5.3.1
rarfile==3.1
rebulk==3.0.1
redis==2.10.5
regex==2020.2.20
reportlab==3.3.0
requests==2.23.0
resampy==0.2.2
rsa==4.0
scenedetect==0.5.1.1
scikit-learn==0.22.2.post1
scipy==1.4.1
selenium==3.141.0
simplejson==3.17.0
six==1.10.0
sklearn==0.0
SoundFile==0.10.3.post1
soupsieve==2.0
splinter==0.13.0
sqlparse==0.2.2
stevedore==1.32.0
subliminal==2.1.0
swapper==1.1.2.post1
tabula-py==2.0.4
termcolor==1.1.0
tqdm==4.43.0
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
vine==1.1.3
watchdog==0.9.0
wavio==0.0.4
wcwidth==0.1.8
wit==6.0.0
xlrd==1.2.0
XlsxWriter==1.2.8
sendgrid
djangorestframework-simplejwt